requirements.txt
lark-parser~=0.12.0
pandas~=2.0.1
PyYAML~=6.0
openpyxl~=3.1.2